requirements.txt
Flask==2.3.2
gunicorn==20.1.0
flask-restx==1.3.0
flask-swagger-ui==4.11.1
Werkzeug==3.0.1
pytest
requests
pyodbc==5.2.0
Flask-Cors==4.0.1
pylint
flask-jwt-extended
psycopg2-binary==2.9.10